refactor: split models into focused sub-modules
Break src/models/mod.rs and src/models/ecs.rs into domain-scoped files:

- event.rs: NormalizedEvent, EventCategory, ProcessContext
- fields.rs: EventFields and all category-specific field structs
- alert.rs: Alert, AlertSeverity, DetectionEngine
- match_details.rs: MatchDetails, Sigma/YARA/IOC detail structs
- ecs/mod.rs: public ECS exports and top-level conversion
- ecs/ sub-modules: alert, event, network, user, registry, helpers

Existing crate::models::* imports unchanged via re-exports.
Serialized alert and ECS output unchanged.
