+++ src/models/ecs/mod.rs
@@ -3,726 +3,29 @@
 //! This module provides a translation layer between internal Alert structures
 //! and the standardized ECS format expected by SIEM systems like Elasticsearch,
 //! Splunk, and other log aggregation platforms.
-//!
-//! ## Design
-//! - Decouples internal data models from external API contracts
-//! - Ensures consistent JSON output for security alerts
-//! - Supports incremental field additions without breaking changes
 
-use crate::models::{
-    Alert, AlertSeverity, EventCategory, EventFields, MatchDetails, ProcessContext,
+mod alert;
+mod event;
+mod helpers;
+mod network;
+mod registry;
+mod user;
+
+pub use alert::{DnsAnswer, EcsAlert};
+
+use crate::models::{Alert, EventFields};
+use event::{
+    alert_severity_to_event_severity, ecs_event_action, ecs_event_category, ecs_event_type,
+    event_dataset, event_provider, host_os_family, host_os_type, network_direction_from_category,
 };
-use crate::sensor::Platform;
-use serde::Serialize;
-use std::net::IpAddr;
+use helpers::{basename, file_extension_from_path, parse_bool, parse_u16, parse_u64};
+use network::{extract_ips, network_transport_from_opcode, network_type_from_ip};
+use registry::split_registry_path;
+use user::apply_user_fields;
 
 const ECS_VERSION: &str = "9.3.0";
 const EVENT_MODULE: &str = "edr";
 
-#[derive(Serialize)]
-pub struct DnsAnswer {
-    #[serde(rename = "data")]
-    pub data: String,
-}
-
-/// ECS-compliant alert structure for SIEM ingestion
-#[derive(Serialize)]
-pub struct EcsAlert {
-    /// Event timestamp in ISO 8601 format
-    #[serde(rename = "@timestamp")]
-    pub timestamp: String,
-
-    /// ECS schema version
-    #[serde(rename = "ecs.version")]
-    pub ecs_version: String,
-
-    /// Event kind (always "alert" for detections)
-    #[serde(rename = "event.kind")]
-    pub event_kind: String,
-
-    /// Event category for classification
-    #[serde(rename = "event.category", skip_serializing_if = "Vec::is_empty")]
-    pub event_category: Vec<String>,
-
-    /// Event type (subcategory)
-    #[serde(rename = "event.type", skip_serializing_if = "Vec::is_empty")]
-    pub event_type: Vec<String>,
-
-    /// Event action
-    #[serde(rename = "event.action", skip_serializing_if = "Option::is_none")]
-    pub event_action: Option<String>,
-
-    /// Event code (e.g. Windows Event ID / Sysmon Event ID)
-    #[serde(rename = "event.code", skip_serializing_if = "Option::is_none")]
-    pub event_code: Option<String>,
-
-    /// Event severity (numeric)
-    #[serde(rename = "event.severity", skip_serializing_if = "Option::is_none")]
-    pub event_severity: Option<u8>,
-
-    /// ECS module (integration name)
-    #[serde(rename = "event.module")]
-    pub event_module: String,
-
-    /// ECS dataset within the module
-    #[serde(rename = "event.dataset")]
-    pub event_dataset: String,
-
-    /// Source that generated the event
-    #[serde(rename = "event.provider")]
-    pub event_provider: String,
-
-    /// Host OS type for the sensor that produced the event.
-    #[serde(rename = "host.os.type")]
-    pub host_os_type: String,
-
-    /// Host OS family for the sensor that produced the event.
-    #[serde(rename = "host.os.family")]
-    pub host_os_family: String,
-
-    /// Detection rule name
-    #[serde(rename = "rule.name")]
-    pub rule_name: String,
-
-    /// Detection rule description / IOC comment
-    #[serde(rename = "rule.description", skip_serializing_if = "Option::is_none")]
-    pub rule_description: Option<String>,
-
-    /// Detection severity (critical, high, medium, low)
-    #[serde(rename = "edr.rule.severity")]
-    pub edr_rule_severity: String,
-
-    /// Detection engine (Sigma, Yara, Ioc)
-    #[serde(rename = "edr.rule.engine")]
-    pub edr_rule_engine: String,
-
-    // ========================================================================
-    // Process Fields
-    // ========================================================================
-    #[serde(rename = "process.executable", skip_serializing_if = "Option::is_none")]
-    pub process_executable: Option<String>,
-
-    #[serde(rename = "process.name", skip_serializing_if = "Option::is_none")]
-    pub process_name: Option<String>,
-
-    #[serde(
-        rename = "process.command_line",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_command_line: Option<String>,
-
-    #[serde(rename = "process.pid", skip_serializing_if = "Option::is_none")]
-    pub process_pid: Option<u64>,
-
-    #[serde(
-        rename = "process.parent.executable",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_parent_executable: Option<String>,
-
-    #[serde(
-        rename = "process.parent.name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_parent_name: Option<String>,
-
-    #[serde(
-        rename = "process.parent.command_line",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_parent_command_line: Option<String>,
-
-    #[serde(rename = "process.parent.pid", skip_serializing_if = "Option::is_none")]
-    pub process_parent_pid: Option<u64>,
-
-    #[serde(
-        rename = "process.working_directory",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_working_directory: Option<String>,
-
-    #[serde(
-        rename = "edr.process.integrity_level",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_process_integrity_level: Option<String>,
-
-    #[serde(
-        rename = "process.pe.original_file_name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_original_file_name: Option<String>,
-
-    #[serde(rename = "process.pe.product", skip_serializing_if = "Option::is_none")]
-    pub process_product: Option<String>,
-
-    #[serde(
-        rename = "process.pe.description",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub process_description: Option<String>,
-
-    #[serde(rename = "user.name", skip_serializing_if = "Option::is_none")]
-    pub user_name: Option<String>,
-
-    #[serde(rename = "user.id", skip_serializing_if = "Option::is_none")]
-    pub user_id: Option<String>,
-
-    #[serde(rename = "user.domain", skip_serializing_if = "Option::is_none")]
-    pub user_domain: Option<String>,
-
-    #[serde(rename = "winlog.logon.id", skip_serializing_if = "Option::is_none")]
-    pub winlog_logon_id: Option<String>,
-
-    #[serde(rename = "winlog.logon.guid", skip_serializing_if = "Option::is_none")]
-    pub winlog_logon_guid: Option<String>,
-
-    // ========================================================================
-    // Network Fields
-    // ========================================================================
-    #[serde(rename = "destination.ip", skip_serializing_if = "Option::is_none")]
-    pub destination_ip: Option<String>,
-
-    #[serde(rename = "destination.port", skip_serializing_if = "Option::is_none")]
-    pub destination_port: Option<u16>,
-
-    #[serde(rename = "source.ip", skip_serializing_if = "Option::is_none")]
-    pub source_ip: Option<String>,
-
-    #[serde(rename = "source.port", skip_serializing_if = "Option::is_none")]
-    pub source_port: Option<u16>,
-
-    #[serde(rename = "destination.domain", skip_serializing_if = "Option::is_none")]
-    pub destination_domain: Option<String>,
-
-    #[serde(rename = "network.transport", skip_serializing_if = "Option::is_none")]
-    pub network_transport: Option<String>,
-
-    #[serde(rename = "network.type", skip_serializing_if = "Option::is_none")]
-    pub network_type: Option<String>,
-
-    #[serde(rename = "network.protocol", skip_serializing_if = "Option::is_none")]
-    pub network_protocol: Option<String>,
-
-    #[serde(rename = "network.direction", skip_serializing_if = "Option::is_none")]
-    pub network_direction: Option<String>,
-
-    // ========================================================================
-    // File Fields
-    // ========================================================================
-    #[serde(rename = "file.path", skip_serializing_if = "Option::is_none")]
-    pub file_path: Option<String>,
-
-    #[serde(rename = "file.name", skip_serializing_if = "Option::is_none")]
-    pub file_name: Option<String>,
-
-    #[serde(rename = "file.extension", skip_serializing_if = "Option::is_none")]
-    pub file_extension: Option<String>,
-
-    #[serde(rename = "file.created", skip_serializing_if = "Option::is_none")]
-    pub file_created: Option<String>,
-
-    #[serde(
-        rename = "edr.file.previous_created",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_file_previous_created: Option<String>,
-
-    #[serde(
-        rename = "file.pe.original_file_name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub file_original_file_name: Option<String>,
-
-    #[serde(rename = "file.pe.product", skip_serializing_if = "Option::is_none")]
-    pub file_product: Option<String>,
-
-    #[serde(
-        rename = "file.pe.description",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub file_description: Option<String>,
-
-    #[serde(
-        rename = "file.code_signature.exists",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub file_code_signature_exists: Option<bool>,
-
-    #[serde(
-        rename = "file.code_signature.subject_name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub file_code_signature_subject_name: Option<String>,
-
-    // ========================================================================
-    // DLL Fields
-    // ========================================================================
-    #[serde(rename = "dll.name", skip_serializing_if = "Option::is_none")]
-    pub dll_name: Option<String>,
-
-    #[serde(rename = "dll.path", skip_serializing_if = "Option::is_none")]
-    pub dll_path: Option<String>,
-
-    // ========================================================================
-    // Registry Fields
-    // ========================================================================
-    #[serde(rename = "registry.path", skip_serializing_if = "Option::is_none")]
-    pub registry_path: Option<String>,
-
-    #[serde(rename = "registry.hive", skip_serializing_if = "Option::is_none")]
-    pub registry_hive: Option<String>,
-
-    #[serde(rename = "registry.key", skip_serializing_if = "Option::is_none")]
-    pub registry_key: Option<String>,
-
-    #[serde(rename = "registry.value", skip_serializing_if = "Option::is_none")]
-    pub registry_value: Option<String>,
-
-    #[serde(
-        rename = "registry.data.strings",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub registry_data_strings: Option<Vec<String>>,
-
-    #[serde(
-        rename = "edr.registry.event_type",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_registry_event_type: Option<String>,
-
-    #[serde(
-        rename = "edr.registry.new_name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_registry_new_name: Option<String>,
-
-    // ========================================================================
-    // DNS Fields
-    // ========================================================================
-    #[serde(rename = "dns.question.name", skip_serializing_if = "Option::is_none")]
-    pub dns_query: Option<String>,
-
-    #[serde(rename = "dns.answers", skip_serializing_if = "Option::is_none")]
-    pub dns_answers: Option<Vec<DnsAnswer>>,
-
-    #[serde(rename = "dns.response_code", skip_serializing_if = "Option::is_none")]
-    pub dns_response_code: Option<String>,
-
-    #[serde(rename = "dns.resolved_ip", skip_serializing_if = "Option::is_none")]
-    pub dns_resolved_ip: Option<Vec<String>>,
-
-    // ========================================================================
-    // Service Persistence Fields
-    // ========================================================================
-    #[serde(rename = "service.name", skip_serializing_if = "Option::is_none")]
-    pub service_name: Option<String>,
-
-    #[serde(
-        rename = "edr.service.executable",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_service_executable: Option<String>,
-
-    #[serde(rename = "edr.service.type", skip_serializing_if = "Option::is_none")]
-    pub edr_service_type: Option<String>,
-
-    #[serde(
-        rename = "edr.service.start_type",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_service_start_type: Option<String>,
-
-    #[serde(
-        rename = "edr.service.account_name",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_service_account_name: Option<String>,
-
-    // ========================================================================
-    // Task Scheduler Persistence Fields
-    // ========================================================================
-    #[serde(rename = "edr.task.name", skip_serializing_if = "Option::is_none")]
-    pub edr_task_name: Option<String>,
-
-    #[serde(rename = "edr.task.content", skip_serializing_if = "Option::is_none")]
-    pub edr_task_content: Option<String>,
-
-    #[serde(rename = "edr.task.user_name", skip_serializing_if = "Option::is_none")]
-    pub edr_task_user_name: Option<String>,
-
-    // ========================================================================
-    // PowerShell Fields
-    // ========================================================================
-    #[serde(
-        rename = "edr.powershell.script_block_text",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_powershell_script_block_text: Option<String>,
-
-    #[serde(
-        rename = "edr.powershell.script_block_id",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_powershell_script_block_id: Option<String>,
-
-    // ========================================================================
-    // WMI Fields
-    // ========================================================================
-    #[serde(rename = "edr.wmi.operation", skip_serializing_if = "Option::is_none")]
-    pub edr_wmi_operation: Option<String>,
-
-    #[serde(rename = "edr.wmi.query", skip_serializing_if = "Option::is_none")]
-    pub edr_wmi_query: Option<String>,
-
-    #[serde(rename = "edr.wmi.namespace", skip_serializing_if = "Option::is_none")]
-    pub edr_wmi_namespace: Option<String>,
-
-    #[serde(rename = "edr.wmi.event_type", skip_serializing_if = "Option::is_none")]
-    pub edr_wmi_event_type: Option<String>,
-
-    // ========================================================================
-    // Remote Thread Fields
-    // ========================================================================
-    #[serde(
-        rename = "edr.remote_thread.target_pid",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_remote_thread_target_pid: Option<u64>,
-
-    #[serde(
-        rename = "edr.remote_thread.target_image",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_remote_thread_target_image: Option<String>,
-
-    #[serde(
-        rename = "edr.remote_thread.start_address",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_remote_thread_start_address: Option<String>,
-
-    #[serde(
-        rename = "edr.remote_thread.start_module",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_remote_thread_start_module: Option<String>,
-
-    #[serde(
-        rename = "edr.remote_thread.start_function",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_remote_thread_start_function: Option<String>,
-
-    // ========================================================================
-    // Process Target Fields
-    // ========================================================================
-    #[serde(
-        rename = "edr.process.target_image",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub edr_process_target_image: Option<String>,
-
-    // ========================================================================
-    // Related Fields
-    // ========================================================================
-    #[serde(rename = "related.ip", skip_serializing_if = "Option::is_none")]
-    pub related_ip: Option<Vec<String>>,
-
-    #[serde(rename = "related.user", skip_serializing_if = "Option::is_none")]
-    pub related_user: Option<Vec<String>>,
-
-    // ========================================================================
-    // Debug Match Details
-    // ========================================================================
-    #[serde(rename = "edr.match", skip_serializing_if = "Option::is_none")]
-    pub edr_match: Option<MatchDetails>,
-}
-
-/// Convert internal Alert to ECS format
-fn alert_severity_to_event_severity(severity: AlertSeverity) -> u8 {
-    match severity {
-        AlertSeverity::Low => 25,
-        AlertSeverity::Medium => 50,
-        AlertSeverity::High => 75,
-        AlertSeverity::Critical => 100,
-    }
-}
-
-fn ecs_event_category(category: EventCategory) -> Vec<String> {
-    match category {
-        EventCategory::Process => vec!["process".to_string()],
-        EventCategory::Network => vec!["network".to_string()],
-        EventCategory::File => vec!["file".to_string()],
-        EventCategory::Registry => vec!["registry".to_string()],
-        EventCategory::Dns => vec!["network".to_string()],
-        EventCategory::ImageLoad => vec!["library".to_string()],
-        EventCategory::Scripting => vec!["process".to_string()],
-        EventCategory::Wmi => vec!["api".to_string()],
-        EventCategory::Service => vec!["configuration".to_string()],
-        EventCategory::Task => vec!["configuration".to_string()],
-    }
-}
-
-fn ecs_event_type(category: EventCategory, opcode: u8, event_id: u16) -> Vec<String> {
-    match category {
-        EventCategory::Process => match opcode {
-            1 => vec!["start".to_string()],
-            2 => vec!["end".to_string()],
-            _ => vec!["info".to_string()],
-        },
-        EventCategory::Network => vec!["connection".to_string()],
-        EventCategory::File => match opcode {
-            64 => vec!["creation".to_string()],
-            70 | 72 => vec!["deletion".to_string()],
-            71 => vec!["change".to_string()],
-            _ => vec!["change".to_string()],
-        },
-        EventCategory::Registry => match opcode {
-            36 => vec!["creation".to_string()],
-            38 | 41 => vec!["deletion".to_string()],
-            39 => vec!["change".to_string()],
-            _ => vec!["change".to_string()],
-        },
-        EventCategory::Dns => vec!["protocol".to_string()],
-        EventCategory::ImageLoad => vec!["start".to_string()],
-        EventCategory::Scripting => vec!["info".to_string()],
-        EventCategory::Wmi => vec!["info".to_string()],
-        EventCategory::Service => {
-            if event_id == 7045 {
-                vec!["creation".to_string()]
-            } else {
-                vec!["change".to_string()]
-            }
-        }
-        EventCategory::Task => {
-            if event_id == 106 {
-                vec!["creation".to_string()]
-            } else {
-                vec!["change".to_string()]
-            }
-        }
-    }
-}
-
-fn ecs_event_action(category: EventCategory, opcode: u8, event_id: u16) -> Option<String> {
-    let action = match category {
-        EventCategory::Process => match opcode {
-            1 => "process-start",
-            2 => "process-end",
-            _ => "process-info",
-        },
-        EventCategory::Network => "network-connection",
-        EventCategory::File => match opcode {
-            64 => "file-create",
-            70 | 72 => "file-delete",
-            71 => "file-rename",
-            _ => "file-change",
-        },
-        EventCategory::Registry => match opcode {
-            36 => "registry-create",
-            38 | 41 => "registry-delete",
-            39 => "registry-set",
-            _ => "registry-change",
-        },
-        EventCategory::Dns => "dns-query",
-        EventCategory::ImageLoad => "image-load",
-        EventCategory::Scripting => "powershell-script",
-        EventCategory::Wmi => "wmi-operation",
-        EventCategory::Service => {
-            if event_id == 7045 {
-                "service-create"
-            } else {
-                "service-change"
-            }
-        }
-        EventCategory::Task => {
-            if event_id == 106 {
-                "task-create"
-            } else {
-                "task-change"
-            }
-        }
-    };
-    Some(action.to_string())
-}
-
-fn event_dataset(category: EventCategory) -> String {
-    let suffix = match category {
-        EventCategory::Process => "process",
-        EventCategory::Network => "network",
-        EventCategory::File => "file",
-        EventCategory::Registry => "registry",
-        EventCategory::Dns => "dns",
-        EventCategory::ImageLoad => "library",
-        EventCategory::Scripting => "scripting",
-        EventCategory::Wmi => "wmi",
-        EventCategory::Service => "service",
-        EventCategory::Task => "task",
-    };
-    format!("{}.{}", EVENT_MODULE, suffix)
-}
-
-fn event_provider(alert: &Alert) -> String {
-    alert.event.provider.clone()
-}
-
-fn host_os_type(platform: Platform) -> String {
-    match platform {
-        Platform::Windows => "windows".to_string(),
-        Platform::Linux => "linux".to_string(),
-    }
-}
-
-fn host_os_family(platform: Platform) -> String {
-    match platform {
-        Platform::Windows => "windows".to_string(),
-        Platform::Linux => "linux".to_string(),
-    }
-}
-
-fn network_direction_from_category(category: EventCategory) -> Option<String> {
-    match category {
-        EventCategory::Network | EventCategory::Dns => Some("egress".to_string()),
-        _ => None,
-    }
-}
-
-fn parse_u64(value: &Option<String>) -> Option<u64> {
-    value.as_ref().and_then(|v| v.trim().parse::<u64>().ok())
-}
-
-fn parse_u16(value: &Option<String>) -> Option<u16> {
-    value.as_ref().and_then(|v| v.trim().parse::<u16>().ok())
-}
-
-fn parse_bool(value: &Option<String>) -> Option<bool> {
-    let normalized = value.as_ref()?.trim().to_ascii_lowercase();
-    match normalized.as_str() {
-        "true" | "signed" | "valid" | "yes" => Some(true),
-        "false" | "unsigned" | "invalid" | "no" => Some(false),
-        _ => None,
-    }
-}
-
-fn basename(path: &str) -> Option<String> {
-    let trimmed = path.trim_matches('"');
-    let name = trimmed.rsplit(['\\', '/']).next().unwrap_or("");
-    if name.is_empty() {
-        None
-    } else {
-        Some(name.to_string())
-    }
-}
-
-fn file_extension_from_path(path: &str) -> Option<String> {
-    let name = basename(path)?;
-    let (_, ext) = name.rsplit_once('.')?;
-    if ext.is_empty() {
-        None
-    } else {
-        Some(ext.to_string())
-    }
-}
-
-fn network_transport_from_opcode(opcode: u8) -> Option<String> {
-    match opcode {
-        12 => Some("tcp".to_string()),
-        15 => Some("udp".to_string()),
-        _ => None,
-    }
-}
-
-fn network_type_from_ip(ip: &str) -> Option<String> {
-    match ip.parse::<IpAddr>() {
-        Ok(IpAddr::V4(_)) => Some("ipv4".to_string()),
-        Ok(IpAddr::V6(_)) => Some("ipv6".to_string()),
-        Err(_) => None,
-    }
-}
-
-fn split_registry_path(
-    path: &str,
-    event_type: Option<&str>,
-) -> (Option<String>, Option<String>, Option<String>) {
-    let mut parts = path.split('\\');
-    let hive = parts.next().map(|value| value.to_string());
-    let rest: Vec<&str> = parts.collect();
-    if rest.is_empty() {
-        return (hive, None, None);
-    }
-
-    let is_value = event_type
-        .map(|value| value.to_ascii_lowercase().contains("value"))
-        .unwrap_or(false);
-
-    if is_value {
-        let value = rest.last().unwrap().to_string();
-        let key = rest[..rest.len() - 1].join("\\");
-        let key = if key.is_empty() { None } else { Some(key) };
-        (hive, key, Some(value))
-    } else {
-        let key = rest.join("\\");
-        let key = if key.is_empty() { None } else { Some(key) };
-        (hive, key, None)
-    }
-}
-
-fn extract_ips(value: &str) -> Vec<String> {
-    let mut ips = Vec::new();
-    for token in value.split(|c: char| !c.is_ascii_hexdigit() && c != '.' && c != ':') {
-        if token.is_empty() {
-            continue;
-        }
-        if let Ok(addr) = token.parse::<IpAddr>() {
-            ips.push(addr.to_string());
-        }
-    }
-    ips.sort();
-    ips.dedup();
-    ips
-}
-
-fn is_sid(value: &str) -> bool {
-    value.starts_with("S-1-")
-}
-
-fn split_user(value: &str) -> (Option<String>, Option<String>, Option<String>) {
-    if is_sid(value) {
-        return (None, Some(value.to_string()), None);
-    }
-
-    if let Some((domain, name)) = value.split_once('\\') {
-        return (Some(name.to_string()), None, Some(domain.to_string()));
-    }
-
-    (Some(value.to_string()), None, None)
-}
-
-fn apply_user_fields(ecs: &mut EcsAlert, user: Option<&str>) {
-    let value = match user {
-        Some(value) if !value.is_empty() => value,
-        _ => return,
-    };
-
-    let (name, id, domain) = split_user(value);
-
-    if ecs.user_id.is_none() {
-        ecs.user_id = id;
-    }
-
-    if ecs.user_domain.is_none() {
-        ecs.user_domain = domain;
-    }
-
-    if let Some(name) = name {
-        if ecs.user_name.is_none() || ecs.user_name.as_deref().map(is_sid).unwrap_or(false) {
-            ecs.user_name = Some(name);
-        }
-    }
-}
-
-/// Convert internal Alert to ECS format
 impl From<&Alert> for EcsAlert {
     fn from(alert: &Alert) -> Self {
         let opcode = alert.event.opcode;
@@ -1032,59 +335,6 @@ impl From<&Alert> for EcsAlert {
         }
 
         ecs
-    }
-}
-
-impl EcsAlert {
-    fn apply_process_context(ecs: &mut EcsAlert, context: &ProcessContext) {
-        if ecs.process_executable.is_none() {
-            ecs.process_executable = context.image.clone();
-        }
-        if ecs.process_command_line.is_none() {
-            ecs.process_command_line = context.command_line.clone();
-        }
-        if ecs.process_pid.is_none() {
-            ecs.process_pid = parse_u64(&context.process_id);
-        }
-        if ecs.process_parent_executable.is_none() {
-            ecs.process_parent_executable = context.parent_image.clone();
-        }
-        if ecs.process_parent_command_line.is_none() {
-            ecs.process_parent_command_line = context.parent_command_line.clone();
-        }
-        if ecs.process_parent_pid.is_none() {
-            ecs.process_parent_pid = parse_u64(&context.parent_process_id);
-        }
-        if ecs.process_working_directory.is_none() {
-            ecs.process_working_directory = context.current_directory.clone();
-        }
-        if ecs.edr_process_integrity_level.is_none() {
-            ecs.edr_process_integrity_level = context.integrity_level.clone();
-        }
-        if ecs.process_original_file_name.is_none() {
-            ecs.process_original_file_name = context.original_file_name.clone();
-        }
-        if ecs.process_product.is_none() {
-            ecs.process_product = context.product.clone();
-        }
-        if ecs.process_description.is_none() {
-            ecs.process_description = context.description.clone();
-        }
-        apply_user_fields(ecs, context.user.as_deref());
-        if ecs.winlog_logon_id.is_none() {
-            ecs.winlog_logon_id = context.logon_id.clone();
-        }
-        if ecs.winlog_logon_guid.is_none() {
-            ecs.winlog_logon_guid = context.logon_guid.clone();
-        }
-
-        if ecs.process_name.is_none() {
-            ecs.process_name = ecs.process_executable.as_deref().and_then(basename);
-        }
-
-        if ecs.process_parent_name.is_none() {
-            ecs.process_parent_name = ecs.process_parent_executable.as_deref().and_then(basename);
-        }
     }
 }
